Update README.md
Updated README to reflect that you no longer need to enable Developer Mode to use WSL,
also provided more modern install instructions - enabling the WSL feature from PowerShell
and reflecting the fact that you can now choose between many distributions not just Ubuntu.

The need for the shortened URL is because GitHub does not support linking to non http(s) URIs
such as `ms-windows-store://search/?query=wsl` which is what we need to do to open the Microsoft Store
on Windows 10 computers to the correct page. A URL shortener that masks the URI-Scheme behind https
works around this limitation.